requirements.txt
line-bot-sdk==3.14.0
fastapi
uvicorn[standard]
google.generativeai
langchain
langchain-google-genai
yfinance
pydantic
tiktoken
Pillow